+++ lib/pdfmonkey/document.rb
@@ -30,15 +30,16 @@ module Pdfmonkey
     attr_reader :attributes
     def_delegators :attributes, *ATTRIBUTES
 
-    def self.generate!(document_template_id, payload)
-      document = generate(document_template_id, payload)
+    def self.generate!(document_template_id, payload, meta = {})
+      document = generate(document_template_id, payload, meta)
       document.reload! until document.done?
       document
     end
 
-    def self.generate(template_id, payload)
+    def self.generate(template_id, payload, meta = {})
       document = new(
         document_template_id: template_id,
+        meta: meta.to_json,
         payload: payload.to_json,
         status: 'pending')